メモリーリークを検知するコードを追加

+++ AppFrame/source/Collision/Plane3D.h
@@ -4,7 +4,7 @@
 class Plane3D {
 public:
 	Plane3D(){}
-	Plane3D(double a, double b, double c, double d){
+	Plane3D(float a, float b, float c, float d){
 		this->a = a;
 		this->b = b;
 		this->c = c;
